+++ src/view/(auth)/sign-up/Index.tsx
